requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorclass==2.2.2
dj-database-url==2.2.0
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-heroku==0.3.1
django-rest-authtoken==2.1.4
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.15.2
dnspython==2.6.1
docopt==0.6.2
drf-yasg==1.21.7
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
mongoengine==0.28.2
packaging==24.1
pillow==10.4.0
pip-upgrader==1.4.15
psycopg2==2.9.10
pymongo==4.8.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.3.0
sqlparse==0.5.0
terminaltables==3.1.10
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.7.0
